paste/requirements.txt
fastapi==0.65.1
pydantic==1.8.2
tortoise-orm[asyncpg]==0.17.3
aioredis==1.3.1
aerich==0.5.3
sonyflake-py==1.2.1
python-multipart==0.0.5